Remove redundant name prefix for libraries.
