+++ Microsoft.MixedReality.Sharing.sln
@@ -3,13 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.645
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.MixedReality.Sharing.Network", "libs\Microsoft.MixedReality.Sharing.Network\src\Microsoft.MixedReality.Sharing.Network.csproj", "{09BB4917-1C3B-4C31-B578-0614286E7EFC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.MixedReality.Sharing.Network", "libs\Network\src\Microsoft.MixedReality.Sharing.Network.csproj", "{09BB4917-1C3B-4C31-B578-0614286E7EFC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.MixedReality.Sharing.StateSync", "libs\Microsoft.MixedReality.Sharing.StateSync\src\Microsoft.MixedReality.Sharing.StateSync.csproj", "{1F906578-3A91-4F95-AD37-282139D65EA8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.MixedReality.Sharing.StateSync", "libs\StateSync\src\Microsoft.MixedReality.Sharing.StateSync.csproj", "{1F906578-3A91-4F95-AD37-282139D65EA8}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.MixedReality.Sharing.StateSync.Test", "libs\Microsoft.MixedReality.Sharing.StateSync\test\Microsoft.MixedReality.Sharing.StateSync.Test.csproj", "{3A6EAB71-070D-4B03-B166-12A8886D0223}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.MixedReality.Sharing.StateSync.Test", "libs\StateSync\test\Microsoft.MixedReality.Sharing.StateSync.Test.csproj", "{3A6EAB71-070D-4B03-B166-12A8886D0223}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.MixedReality.Sharing.Matchmaking", "libs\Microsoft.MixedReality.Sharing.Matchmaking\src\Microsoft.MixedReality.Sharing.Matchmaking.csproj", "{CC5B96F6-669E-4BAD-98E1-072801F771AA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.MixedReality.Sharing.Matchmaking", "libs\Matchmaking\src\Microsoft.MixedReality.Sharing.Matchmaking.csproj", "{CC5B96F6-669E-4BAD-98E1-072801F771AA}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "TempLib", "libs\TempLib\src\TempLib.vcxproj", "{B69106CA-ECD6-49CC-A1A1-E5D97E9EB9E0}"
 EndProject
