+++ libs/Matchmaking/src/IRoomInfo.cs
