+++ libs/Matchmaking/src/IRoomManager.cs
