--- libs/Microsoft.MixedReality.Sharing.StateSync/test/UnitTest1.cs
@@ -1,24 +1,0 @@
-using System;
-using Xunit;
-
-namespace Microsoft.MixedReality.Sharing.StateSync.Test
-{
-    public class UnitTest1
-    {
-        [Fact]
-        public void Test1()
-        {
-            var snap = new Snapshot();
-            Assert.NotNull(snap.GetType());
-        }
-
-        [Theory]
-        [InlineData("p")]
-        [InlineData("O;Malley")]
-        [InlineData("O,Malley")]
-        public void TestWithHardcodedParams(string name)
-        {
-            Assert.Equal(-1, name.IndexOf('x'));
-        }
-    }
-}
