+++ libs/Network/src/IChannel.cs
