+++ libs/Network/src/INetParticipant.cs
