+++ libs/Network/src/ISession.cs
