+++ libs/Network/src/Microsoft.MixedReality.Sharing.Network.csproj
