+++ libs/StateSync/src/ISynchronizationStore.cs
