+++ libs/StateSync/src/ITransaction.cs
