+++ libs/StateSync/src/Microsoft.MixedReality.Sharing.StateSync.csproj
