+++ libs/StateSync/src/Snapshot.cs
